[feat] 팀 빌딩 기능 수정
